finished my backend

+++ routes/profiles.js
@@ -1,5 +1,7 @@
 const express = require("express");
 const Router = express.Router();
+const request = require("request");
+const config = require("config");
 
 //load auth token middleware
 const auth = require("./../middleware/auth");
@@ -159,6 +161,10 @@ Router.get("/user/:user_id", async (req, res) => {
 */
 Router.delete("/", auth, async (req, res) => {
   try {
+    let user = await User.findOne({ user: req.user.id });
+
+    if (!user) return res.status(400).json({ mdg: "There are no user" });
+
     await Profile.findOneAndRemove({ user: req.user.id });
     await User.findOneAndRemove({ _id: req.user.id });
     res.json({ msg: "User deleted" });
@@ -166,6 +172,197 @@ Router.delete("/", auth, async (req, res) => {
     console.log(err.message);
     res.status(500).json({ msg: "Server Error" });
   }
+});
+
+/*
+@route  PUT api/profile/experience
+@desc   Update profile experience
+@access private
+*/
+Router.put(
+  "/experience",
+  [
+    auth,
+    check("title", "Title is required")
+      .not()
+      .isEmpty(),
+    check("company", "Company is required")
+      .not()
+      .isEmpty(),
+    check("from", "From date is required")
+      .not()
+      .isEmpty()
+  ],
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty())
+      return res.status(400).json({ errors: errors.array() });
+
+    const {
+      title,
+      company,
+      from,
+      current,
+      description,
+      location,
+      to
+    } = req.body;
+
+    const newExp = {
+      title,
+      company,
+      from,
+      current,
+      description,
+      location,
+      to
+    };
+
+    try {
+      const profile = await Profile.findOne({ user: req.user.id });
+      profile.experience.unshift(newExp);
+      await profile.save();
+      res.json(profile);
+    } catch (err) {
+      console.log(err.message);
+      res.status(500).send("Server Error");
+    }
+  }
+);
+
+/*
+@route  DELETE api/profile/experience/:exp_id
+@desc   delete profile experience
+@access private
+*/
+Router.delete("/experience/:exp_id", auth, async (req, res) => {
+  try {
+    const profile = await Profile.findOne({ user: req.user.id });
+
+    const getIndex = profile.experience
+      .map(item => item.id)
+      .indexOf(req.params.exp_id);
+
+    if (getIndex === -1)
+      return res
+        .status(400)
+        .json({ msg: "There are no experience for this id" });
+
+    profile.experience.splice(getIndex, 1);
+    await profile.save();
+    res.json(profile);
+  } catch (err) {
+    console.log(err.message);
+    res.status(500).send("Server Error");
+  }
+});
+
+/*
+@route  PUT api/profile/education/:edu_id
+@desc   Update profile education
+@access private
+*/
+Router.put(
+  "/education",
+  [
+    auth,
+    check("school", "School is required")
+      .not()
+      .isEmpty(),
+    check("degree", "Degree is required")
+      .not()
+      .isEmpty(),
+    check("fieldofstudy", "Field of study is required")
+      .not()
+      .isEmpty()
+  ],
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty())
+      return res.status(400).json({ errors: errors.array() });
+
+    const {
+      school,
+      degree,
+      fieldofstudy,
+      from,
+      to,
+      current,
+      description
+    } = req.body;
+
+    const newEdu = {
+      school,
+      degree,
+      fieldofstudy,
+      from,
+      to,
+      current,
+      description
+    };
+
+    try {
+      const profile = await Profile.findOne({ user: req.user.id });
+      profile.education.unshift(newEdu);
+      await profile.save();
+      res.json(profile);
+    } catch (err) {
+      console.log(err);
+      res.status(500).send("Server Error");
+    }
+  }
+);
+
+/*
+@route  DELETE api/profile/education/:edu_id
+@desc   delete profile education
+@access private
+*/
+Router.delete("/education/:edu_id", auth, async (req, res) => {
+  try {
+    const profile = await Profile.findOne({ user: req.user.id });
+    const getIndex = profile.education
+      .map(item => item.id)
+      .indexOf(req.params.edu_id);
+
+    if (getIndex === -1)
+      return res
+        .status(400)
+        .json({ msg: "There are no education for this profile" });
+
+    profile.education.splice(getIndex, 1);
+    await profile.save();
+    res.json(profile);
+  } catch (err) {
+    console.log(err.message);
+    res.status(500).send("Server Error");
+  }
+});
+
+/*
+@route  GET api/profile/github/:username
+@desc   GET profile github username
+@access public
+*/
+Router.get("/github/:username", (req, res) => {
+  const options = {
+    uri: `https://api.github.com/users/${
+      req.params.username
+    }/repos?per_page=5&sort=created:asc&client_id=${config.get(
+      "githubClientId"
+    )}&client_secret=${config.get("githubClientSecret")}`,
+    method: "GET",
+    headers: { "user-agent": "node.js" }
+  };
+
+  request(options, (err, response, body) => {
+    if (err) console.log(err);
+
+    if (response.statusCode !== 200)
+      return res.status(400).json({ msg: "No Github Profile Found" });
+
+    res.json(JSON.parse(body));
+  });
 });
 
 module.exports = Router;
